NIGHTMARE COMMIT: The original plan for this commit was to add the ability for a logged-in user to create a new thread with an initial post, which it does but poorly.
Two bugs were encountered trying to make this happen.

1. It was discovered that the .query method of the models uses a different session than the one created by db.session. When a model instance that was created by a query is added via session.add(), and exception is thrown. This is the case of the authenticated_user object when associated with a newly created Thread or Post.

In the end, the hackish fix for this problem was to add an additional constructor argument to the affected models to take integer values of foreign keys, rather than the related objects which are attached to different sessions. This does not follow the pattern seen elsewhere for creating models, and it seems that there should be a better way to configure SQLAlchemy so that all interactions with the database on a given request use a common session.

2. After creating a workaround for thre first bug, new rows could be successfully created in the database, but the view on the homepage that displays threads does not show the new threads, even upon refresh. The only way to get the new data to display is to restart the application. I do not know how to fix this problem. I imagine that there is a SQLAlchemy configuration that I am unaware of that would fix this problem.